tests: refactor and split up
Refactor the singular test file into multiple for slightly better
compile times in the future and also more manageable workloads.
